sheets/qrcode/requirements.txt
qrcode==7.4.2
pillow==10.2.0
matplotlib==3.8.2 